Created Resource.configure. There is a lot of duplication on Resource.Update however it is not as bad as it seems. Agda is planning on adding language support for updating records with a particular field. This would eliminate the current duplication.
An alternative to get a DRY
implementation without waiting for Agda updates is to use a
PartialConfig whose type signature would be the same as the one
appearing in Resource.configure (List (Σ Code convert)). Then modules
using a config would take a TotalConfig with an extra parameter proof
of totality (All resource hooks being keys in the list of products).

The benefit we get by keeping Config a record is that its
construction is guaranteed to be total, so no extra proof is necessary.